refactor: filters that are integrated in header comlums
